Automate network config
A lot of the network config can be trivially read from NVRAM.
With a bit of shell scripting, we can extract all the info we need.
This means that config stuff should be hands-off within a model family,
though manually changes are likely needed to LAN_IFACE, WLAN_IFACE
and LAN_BR on models configured differently.
This also means that network settings can be changed from the GUI
without breaking these scripts.

Signed-off-by: ericloewe <ericloewe@outlook.pt>